fixed budgetform redirect, change graphql values from int => float, deleted old files, and cleaned up root folder

+++ server/schemas/typeDefs.js
@@ -26,7 +26,7 @@ type Expense {
   _id: ID
   name: String
   date: String
-  cost: Int
+  cost: Float
   category: Category
 }
 
@@ -34,7 +34,7 @@ type Income {
   _id: ID
   name: String
   date: String
-  amount: Int
+  amount: Float
 }
 
 type Auth {
@@ -46,14 +46,14 @@ type Mutation {
   addUser(firstName: String!, lastName: String!, email: String!, password: String!): Auth
   login(email: String!, password: String!): Auth
   addBudget(userId: ID! budgetMonth: String! budgetYear: Int!): User
-  addExpense(budgetId: ID!, name: String!, cost: Int!, categoryId: ID!): Budget
-  addIncome(budgetId: ID!, name: String!, amount: Int!): Budget
+  addExpense(budgetId: ID!, name: String!, cost: Float!, categoryId: ID!): Budget
+  addIncome(budgetId: ID!, name: String!, amount: Float!): Budget
   removeBudget(budgetId: ID!): Budget
   removeExpense(budgetId: ID!, expenseId: ID!): Budget
   removeIncome(budgetId: ID!, incomeId: ID!): Budget
   updateBudget(budgetId: ID!, budgetMonth: String budgetYear: Int): Budget
-  updateExpense(budgetId: ID!, expenseId: ID!, name: String, cost: Int, categoryId: ID): Budget
-  updateIncome(budgetId: ID!, incomeId: ID!,  name: String, amount: Int): Budget
+  updateExpense(budgetId: ID!, expenseId: ID!, name: String, cost: Float, categoryId: ID): Budget
+  updateIncome(budgetId: ID!, incomeId: ID!,  name: String, amount: Float): Budget
 }
 
 type Query {
